Outrelix/backend/requirements_minimal.txt
fastapi==0.68.2
uvicorn==0.15.0
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==1.8.2
httpx==0.24.1
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
